% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbq-package.R
\docType{package}
\name{cbq-package}
\alias{cbq-package}
\alias{cbq}
\title{The 'cbq' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
