% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{coef.cbq}
\alias{coef.cbq}
\title{Extract CBQ Coefficients}
\usage{
\method{coef}{cbq}(object, ...)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A table of coefficients with their corresponding lower and upper bounds.
}
\description{
Create a table of coefficient results from a \code{cbq} object.
}
