% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbq.R
\name{is.dichotomous}
\alias{is.dichotomous}
\title{Check if a predictor is dichotomous, adopted from package \code{circGLM}}
\usage{
is.dichotomous(x)
}
\arguments{
\item{x}{A character or numerical vector to be tested.}
}
\value{
A logical, \code{TRUE} if the \code{x} has dummy coding (0, 1),
  \code{FALSE} otherwise.
}
\description{
Check if a predictor is dichotomous, adopted from package \code{circGLM}
}
