% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot.cbq}
\alias{plot.cbq}
\title{Plot cbq object}
\usage{
\method{plot}{cbq}(x, type = "trace", ...)
}
\arguments{
\item{x}{A \code{cbq} object to be plotted.}

\item{type}{Character string giving the type of plotting. The options are
\code{"trace"} for trace plots, \code{"coef"} for coefficient plots. The default is the traceplot.}

\item{...}{Additional arguments to be passed to subsequent plot functions.}
}
\value{
None.
}
\description{
General plot function for \code{cbq} objects, which dispatches the chosen
type of plotting to the corresponding function.
}
