% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_coef.cbq}
\alias{plot_coef.cbq}
\title{Make coefficient plots for cbq}
\usage{
plot_coef.cbq(object, ...)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{...}{Additional parameters to be passed to the plot function.}
}
\value{
None.
}
\description{
Plot traceplots from a \code{cbq} object.
}
