% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_trace.cbq}
\alias{plot_trace.cbq}
\title{Make traceplots for cbq}
\usage{
plot_trace.cbq(object, ...)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{...}{Additional parameters to be passed to the traceplot function.}
}
\value{
None.
}
\description{
Plot traceplots from a \code{cbq} object.
}
