% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{predict.cbq}
\alias{predict.cbq}
\title{Predictions based on the fitted parameter values}
\usage{
\method{predict}{cbq}(object, data, ci = 0.95, ...)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{data}{Data used for prediction.}

\item{ci}{Confidence interval. The default is 0.95.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A vector of predictions.
}
\description{
Create a vector of predictions from a \code{cbq} object.
}
