% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print.cbq}
\alias{print.cbq}
\title{Print cbq object}
\usage{
\method{print}{cbq}(object, type = "text", ...)
}
\arguments{
\item{object}{A \code{cbq} object to be printed.}

\item{type}{Character string giving the type of printing, such as
\code{"text"}, \code{"mcmc"}, \code{"coef"}.}

\item{...}{Additional arguments to be passed to print functions.}
}
\description{
General print function for \code{cbq} objects, which dispatches the chosen type
of printing to the corresponding function.
}
\examples{


}
