% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_coef.cbq}
\alias{print_coef.cbq}
\title{Print cbq coefficients}
\usage{
print_coef.cbq(object, digits = 3)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{digits}{Number of digits to display.}
}
\value{
None.
}
\description{
Print cbq coefficients
}
