% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_mcmc.cbq}
\alias{print_mcmc.cbq}
\title{Print the mcmc results from a cbq object}
\usage{
print_mcmc.cbq(object, ...)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{...}{Additional arguments to be passed to the print function.}
}
\value{
None.
}
\description{
This prints a number of diagnostics about the results of a \code{cbq} objects
}
