% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_text.cbq}
\alias{print_text.cbq}
\title{Print the main results from a \code{cbq} object.}
\usage{
print_text.cbq(object, digits = 3)
}
\arguments{
\item{object}{A \code{cbq} object.}

\item{digits}{Number of digits to display.}
}
\description{
Print the main results from a \code{cbq} object.
}
